pyproject.toml
[project]
name = "personal-claude-assistant"
version = "0.1.0"
description = "Personal AI assistant — Telegram bot powered by Claude"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[job-queue]==21.7",
    "anthropic>=0.40.0",
    "ddgs>=6.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
package = false
